client_service/requirements.txt
psycopg2-binary==2.9.10
requests==2.32.3
polars==1.27.1